refactor: normalize dto hydration recursion
